fix hmr
