+++ libs/zzz/formula-ui/src/components/index.ts
@@ -1,1 +1,2 @@
 export * from './TagDisplay'
+export * from './qtMap'
