+++ libs/zzz/formula-ui/src/components/qtMap.ts
@@ -1,0 +1,7 @@
+// TODO: translation
+export const qtMap = {
+  initial: 'Initial',
+  combat: 'Combat',
+  final: 'Final',
+  base: 'Base',
+}
